% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{derfinderPlot-defunct}
\alias{derfinderPlot-defunct}
\alias{plot_cluster}
\alias{plot_overview}
\alias{plot_region_coverage}
\title{Defunct functions in package \sQuote{derfinderPlot}}
\usage{
plot_overview()

plot_cluster()

plot_region_coverage()
}
\description{
These functions are defunct and no longer available.
}
\details{
The following functions are defunct; use the replacement indicated below:
 \itemize{

   \item{plot_overview: \code{\link{plotOverview}}}
   \item{plot_cluster: \code{\link{plotCluster}}}
   \item{plot_region_coverage: \code{\link{plotRegionCoverage}}}

 }
}

