% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derfinderPlot-package.R
\docType{package}
\name{derfinderPlot-package}
\alias{derfinderPlot-package}
\title{Plotting functions for derfinder}
\description{
Plotting functions for derfinder results
}
\author{
Leonardo Collado-Torres \href{mailto:lcolladotor@gmail.com}{lcolladotor@gmail.com}
}
\keyword{package}
