% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Deprecated functions in package \sQuote{derfinderPlot}}
\usage{
plot_cluster()
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{derfinderPlot} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
 the replacement indicated below:
 \itemize{

   \item{plot_overview: \code{\link{plotOverview}}}
   \item{plot_cluster: \code{\link{plotCluster}}}
   \item{plot_region_coverage: \code{\link{plotRegionCoverage}}}

 }
}

