% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCluster}
\alias{plotCluster}
\title{Plot the coverage information surrounding a region cluster}
\usage{
plotCluster(idx, regions, annotation, coverageInfo, groupInfo,
  titleUse = "qval", txdb = NULL, p.ideogram = NULL, ...)
}
\arguments{
\item{idx}{A integer specifying the index number of the region of interest.
This region is graphically highlighted by a red bar.}

\item{regions}{The \code{$regions} output from
\link[derfinder]{calculatePvalues}.}

\item{annotation}{The output from running \link[bumphunter]{annotateNearest}
on the output from \link[derfinder]{calculatePvalues}.}

\item{coverageInfo}{A DataFrame resulting from
\link[derfinder]{loadCoverage} using \code{cutoff=NULL}.}

\item{groupInfo}{A factor specifying the group membership of each sample. It
will be used to color the samples by group.}

\item{titleUse}{Whether to show the p-value (\code{pval}) or the q-value
(\code{qval}) in the title. If \code{titleUse=none} then no p-value or
q-value information is used; useful if no permutations were performed and
thus p-value and q-value information is absent.}

\item{txdb}{A transcript data base such as
\code{TxDb.Hsapiens.UCSC.hg19.knownGene}. If \code{NULL} then no annotation
information is used.}

\item{p.ideogram}{If \code{NULL}, the ideogram for hg19 is built for the
corresponding chromosome. Otherwise an ideogram resuling from
\link[ggbio]{plotIdeogram}.}

\item{...}{Arguments passed to other methods and/or advanced arguments.}
}
\value{
A ggplot2 plot that is ready to be printed out. Tecnically it is a
ggbio object. The region with the red bar is the one whose information is
shown in the title.
}
\description{
For a given region found in \link[derfinder]{calculatePvalues}, plot the
coverage for the cluster this region belongs to as well as some padding. The
mean by group is shown to facilitate comparisons between groups. If
annotation exists, you can plot the trancripts and exons (if any)
overlapping in the vicinity of the region of interest.
}
\details{
See the parameter \code{significantCut} in
\link[derfinder]{calculatePvalues} for how the significance cutoffs are
determined.
}
\examples{
## Load data
library('derfinder')

## Annotate the results
suppressMessages(library('bumphunter'))
annotation <- annotateNearest(genomeRegions$regions, 'hg19')

## Make the plot
suppressMessages(library('TxDb.Hsapiens.UCSC.hg19.knownGene'))
plotCluster(idx=1, regions=genomeRegions$regions, annotation=annotation,
    coverageInfo=genomeDataRaw$coverage, groupInfo=genomeInfo$pop,
    txdb=TxDb.Hsapiens.UCSC.hg19.knownGene)
## Resize the plot window and the labels will look good.

\dontrun{
## For a custom plot, check the ggbio and ggplot2 packages.
## Also feel free to look at the code for this function:
plotCluster

}
}
\author{
Leonardo Collado-Torres
}
\seealso{
\link[derfinder]{loadCoverage}, \link[derfinder]{calculatePvalues},
\link[bumphunter]{annotateNearest}, \link[ggbio]{plotIdeogram}
}

