% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tMatrix.R
\name{tMatrix}
\alias{tMatrix}
\title{Interaction plot between bins of the genome}
\usage{
tMatrix(regions, tilewidth = 1e+06, valueVar = "area", scale = "bp",
  title = "t-Statistic", binPad = 2, palette = SushiColors(7))
}
\arguments{
\item{regions}{A \code{GRanges} object that has a column with the name
specified in \code{valueVar}.}

\item{tilewidth}{A single value that specifies the bin size. Regions
overlapping a bin are used to compare against another mean. Their values
(defined by \code{valueVar}) are used to construct the t-Statistics. Avoid
choosing a small value because that will greatly increase the number of bins
needed. Unless the genome is small.}

\item{valueVar}{This should be a character vector of length one that
specifies which variable from \code{regions} to use to construct the
t-statistic matrix.}

\item{scale}{This argument is passed to \link[Sushi]{labelgenome} and should
be either \code{'bp'}, \code{'Kb'} or \code{'Mb'}.}

\item{title}{This argument is passed to \link[Sushi]{addlegend}.}

\item{binPad}{Defines how many bins to include before and after the first and
last bins that overlap regions.}

\item{palette}{This argument is passed to \link[Sushi]{plotHic} and
\link[Sushi]{addlegend}.}
}
\value{
A list with one element per chromosome. For each chromosome
\code{tMatrix} returns the matrix of t-Statistics supplied to
\link[Sushi]{plotHic}. This function will also make the 'interaction' plots,
one per chromosome.
}
\description{
This function takes a \code{GRanges} object with a variable specifying some
values of interest. It then construct bins for the genome and for each bin
it identifies which regions overlap it. It then calculates t-Statistics
between bins and displays them as an interaction plot. Note that this type
of plot is normally used to visualize HiC data as is done in
\link[Sushi]{plotHic}.
}
\details{
This plot gives an overview of which bins of the genome have similar
mean values from \code{valueVar}. Bins that have 1 or no observations are
assigned a t-Statistics of 0 because \link[Sushi]{plotHic} does not handle
NA values. While these zeros are indistinguable from bins that have the same
means, the reality is that we expect few such cases to happen.

The numerator of the t-Statistics shown in the plot is given by the mean of #' the first bin (leftmost of the two bins) minus the mean of the second bin.
Diagonal elements are z-Scores.
}
\examples{
## Load data
library('derfinder')

## Make plot using a small tilewidth given the small size of the data
res <- tMatrix(genomeRegions$regions, tilewidth = 1e4)

## Change the color palette with grDevices::colorRampPalette()
library('grDevices')
res2 <- tMatrix(genomeRegions$regions, tilewidth = 1e4, palette =
   colorRampPalette(c('grey', 'blue')))

## Now also use a divergin palette from RColorBrewer
library('RColorBrewer')
res3 <- tMatrix(genomeRegions$regions, tilewidth = 1e4,
   palette = colorRampPalette(brewer.pal(11, 'PuOr')))
## If you only have negative or positive t-Statistics, consider using a
## sequential palette such as 'YlGnBu'.
}
\author{
Leonardo Collado-Torres
}
\seealso{
\link[Sushi]{plotHic}, \link[grDevices]{colorRampPalette},
    \link[RColorBrewer]{brewer.pal}
}

