% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{derfinderPlot-package}
\alias{derfinderPlot-package}
\title{Plotting functions for derfinder}
\description{
Plotting functions for derfinder results
}
\author{
Leonardo Collado-Torres <lcollado@jhu.edu>
}
\keyword{package}

